✨ Feat: toast type 설정 및 기본 message 생성
