--- src/types/toastOptions.ts
@@ -1,12 +1,0 @@
-import { ToastOptions } from 'react-toastify';
-
-export const toastOptions: ToastOptions = {
-  position: 'top-center',
-  autoClose: 2000,
-  hideProgressBar: true,
-  closeOnClick: true,
-  pauseOnHover: true,
-  draggable: true,
-  pauseOnFocusLoss: true,
-  closeButton: false,
-};
